pyproject.toml
[project]
name = "lightlab"
version = "0.1.0"
description = "Lightning NLP lab, python 3.10"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "cython>=3.2.4",
    "datasets>=4.7.0",
    "flash-attn",
    "huggingface-hub>=1.6.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "joblib>=1.5.3",
    "lightning>=2.6.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.1",
    "numpy>=2.2.6",
    "optuna==2.10.1",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.1",
    "spacy>=3.8.11",
    "sqlalchemy==1.4.46",
    "torch==2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]


# for if flash attn is ever used
#[tool.uv.extra-build-dependencies]
      #flash-attn = ["torch"]
# prebuilt wheel to avoid massive build, must select a url according to HW/SW specs
#[tool.uv.sources]
#flash-attn = { url = "#" }
